Create abstract SarangiClassifier that implements Classifier and is extended by SarangiSVM, SarangiANN...
