+++ App/src/main/java/com/sarangi/app/App.java
@@ -15,6 +15,7 @@ import com.sarangi.learningmodel.ann.SarangiANN;
 import com.sarangi.learningmodel.ann.ANNRunner;
 import com.sarangi.learningmodel.svm.SarangiSVM;
 import com.sarangi.learningmodel.svm.SVMRunner;
+import com.sarangi.learningmodel.*;
 
 /**
  * A main class for interfacing all the other sub-classes.
@@ -51,18 +52,18 @@ public class App
                 String testFilename = "src/resources/song/songFeatures/test.txt";
 
                 
-               FeatureExtractor.extractFeature(new String("src/resources/song/songFeatures/features.txt"),new String("src/resources/song/Mood_training"));
-               FeatureExtractor.extractFeature(new String("src/resources/song/songFeatures/test.txt"),new String("src/resources/song/Mood_testing"));
+               //FeatureExtractor.extractFeature(new String("src/resources/song/songFeatures/features.txt"),new String("src/resources/song/Mood_training"));
+               //FeatureExtractor.extractFeature(new String("src/resources/song/songFeatures/test.txt"),new String("src/resources/song/Mood_testing"));
 
                 System.out.println("Running SVM...");
-                //SVMRunner svmRunner= new SVMRunner(new String[]{"classical","hiphop","jazz","pop","rock"});
-                SVMRunner svmRunner= new SVMRunner(new String[]{"high_arousal","low_arousal"});
-                svmRunner.run(trainingFilename,testFilename);
+                SVMRunner svmRunner= new SVMRunner(new String[]{"classical","hiphop","jazz","pop","rock"});
+                //SVMRunner svmRunner= new SVMRunner(new String[]{"high_arousal","low_arousal"});
+                svmRunner.run(trainingFilename,testFilename, DatasetUtil.FeatureType.SARANGI_MFCC);
 
                 System.out.println("Running ANN...");
-                //ANNRunner annRunner= new ANNRunner(new String[]{"classical","hiphop","jazz","pop","rock"});
-                ANNRunner annRunner= new ANNRunner(new String[]{"high_arousal","low_arousal"});
-                annRunner.run(trainingFilename,testFilename);
+                ANNRunner annRunner= new ANNRunner(new String[]{"classical","hiphop","jazz","pop","rock"});
+                //ANNRunner annRunner= new ANNRunner(new String[]{"high_arousal","low_arousal"});
+                annRunner.run(trainingFilename,testFilename, DatasetUtil.FeatureType.SARANGI_MFCC);
 
         }
 }
