+++ App/src/main/java/com/sarangi/learningmodel/ann/ANNRunner.java
@@ -39,7 +39,7 @@ public class ANNRunner extends ClassifierRunner {
                 super(labels);
         }
 
-        public void run(String trainingFilename, String testFilename) throws FileNotFoundException, IOException  {
+        public void run(String trainingFilename, String testFilename, DatasetUtil.FeatureType featureType) throws FileNotFoundException, IOException  {
 
                 SongHandler trainingSongHandler = new SongHandler(trainingFilename);
                 List<Song> trainingSongs = trainingSongHandler.loadSongs();
@@ -47,7 +47,7 @@ public class ANNRunner extends ClassifierRunner {
                 SongHandler testSongHandler = new SongHandler(testFilename);
                 List<Song> testSongs = testSongHandler.loadSongs();
 
-                SarangiANN ann = new SarangiANN(trainingSongs,this.labels,DatasetUtil.FeatureType.SARANGI_PITCH);
+                SarangiANN ann = new SarangiANN(trainingSongs,this.labels,featureType);
                 
                 Result result = ann.test(testSongs);
